feat: Add additional track posistion and residual plots (#740)
Added a method to write the track positions and directions that were recorded whenever the navigation has reached a surface to csv. The csv file is read back into the python plotting tools where 2D track position and residual plots are prepared from them (e.g. host vs. device). I moved the csv IO functionality in python into the same files where the corresponding plotting functions are defined.

Also fixed the selection of the number of tracks to check in the validation tools and made sure that the python code is handling the data in double precision, so that the minute residuals between host and device code become visible at all (but only when removing the fast math option for CUDA and even then, it is quantized to multiples of the machine epsilon)

+++ tests/validation/python/plotting/plot_helpers.py
@@ -5,6 +5,7 @@
 # Mozilla Public License Version 2.0
 
 from collections import namedtuple
+import numpy as np
 
 #-------------------------------------------------------------------------------
 # Common helpers for plotting measurement data
@@ -17,13 +18,13 @@ def filter_data(data, filter = lambda df: [], variables = []):
     # Get global data
     if len(filter(data)) == 0:
         for var in variables:
-            dataColl.append(data[var].to_numpy())
+            dataColl.append(data[var].to_numpy(dtype = np.double))
 
     # Filtered data
     else:
         filtered = data.loc[filter]
         for var in variables:
-            dataColl.append(filtered[var].to_numpy())
+            dataColl.append(filtered[var].to_numpy(dtype = np.double))
 
     return tuple(dataColl)
